Create new view extending from day view.
We are not doing anything significant here: this is only to create a new
scheduler view. Over this view, we will add other changes.

Now, why extending day view? Well, this is what the week view does, and
I'm using it as an example. Let's see what is going to happen.