Let's make a column appear for each resource.
Y.SchedulerDayView._valueColDaysNode() is a method to return all columns to be
rendered. By default, it count the number of days to define how many columns to
display. (The Y.SchedulerWeekView does not do much more than setting the number
of days to 7, for example.) You can see this method at
http://tinyurl.com/m3tc3x9

We are going to override this method, so we add a prototype to our resource
view. Instead of counting days, we are going to count resources. The easiest
way to customize it is to copy&paste the original method and change some of its
details. We did it.

Since the copied method depends on many of the locally defined constants from
aui-scheduler-view-day.js (see http://tinyurl.com/mjf2k3t) we are going to copy
and paste the needed ones as well.

Once we do it, we will have three columns - but two of them will be empty:
http://tinyurl.com/n296aj9