Expanding the grid container.
Our previous change was limited because we created new columns, but did not
extend the grid container (which is responsible to get clicks, show rows etc.)
This grid container is extended at Y.SchedulerDayView.syncUI() (cf.
http://tinyurl.com/n58x55x), again taking into account the number of days.

Now, syncUI() does a lot more of things, so we do not want to replace it all.
What we do instead is to

 * override it;
 * call the superclass method over the current instance; and
 * update the grid container colspan.

See the result at http://tinyurl.com/k39wsd5