Adding headers to each column.
Our columns have only one header and we want each one to has its own. The
method responsible to create methods is
Y.SchedulerDayView._valueColHeaderDaysNode() (cf http://tinyurl.com/kceyztg).
So again we will override it, changing it to show a header for each resource.

It will result in the right number of headers, but the headers will show
different days (http://tinyurl.com/l9kxyk7), and we do not want that.