Now we override plotEvents() to plot events in different columns.
Y.SchedulerDayView.plotEvents() (http://tinyurl.com/mm458gy) is the method
responsible for, well, plotting the events. In principle, we just need to
change a few lines - instead of getting the shim by date, get it by resource -
but in practice the method is somewhat monolithic. Sadly, we will have to
copy&paste again.

Now, take a look at the result: http://tinyurl.com/of5q9wv

This change did not work - the events keep appearing only a the first column.
Yet, you can note another interesting thing: events with the same resource
(the pink ones) are properly aligned when intersecting, while the two others
aren't. That's because plotEvents() only handles part of the plotting -
deciding which events are going to be plotting, where to plot and how to align
events in the same "shim"[1]. The two pink events are supposed to be in the
same shim and then plotEvents() can properly align them

The event plotting is actually done elsewhere but, before getting there, we
will need an auxiliary method.

[1] More about it in the next commit.