Creating method for retrieving shim by resource index.
The "shim" is the place where the event is effectively plotted, and the rows
are displayed. Each column has its own shim, and they are listed in the
Y.SchedulerDayView.columnShims attribute.

We need a method that retrieves the shim based on the index of the resource. If
the event resource is the first one, the method will return the shim of the
first column, and so on. Here comes the implementation.